requirements.txt
sounddevice == 0.5.1
pvporcupine == 3.0.3
pvrecorder == 1.2.4
pvcobra == 2.0.3
numpy == 1.26.3
soundfile == 0.12.1
transformers == 4.47.0
openai-whisper == 20240930
requests == 2.32.3
torch == 2.5.1+cu124 
torchaudio == 2.5.1+cu124
keyboard == 0.13.5
Telethon==1.36.0
langchain==0.3.23
duckduckgo_search==2025.4.4
python-dotenv==1.0.1
faster-whisper==1.1.1
colorama==0.4.6
